# Declare the default validators as an empty tuple for faster copying