@thedrow 2015-08-18
It seems that it will break existing code that relies on the validators being a list. I don't think it's worth it.